+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/AdMessageCodec.java
@@ -41,7 +41,6 @@ import java.util.Map;
  * Encodes and decodes values by reading from a ByteBuffer and writing to a ByteArrayOutputStream.
  */
 class AdMessageCodec extends StandardMessageCodec {
-
   // The type values below must be consistent for each platform.
   private static final byte VALUE_AD_SIZE = (byte) 128;
   private static final byte VALUE_AD_REQUEST = (byte) 129;
@@ -71,7 +70,11 @@ class AdMessageCodec extends StandardMessageCodec {
 
   @NonNull Context context;
   @NonNull final FlutterAdSize.AdSizeFactory adSizeFactory;
-  @Nullable private MediationNetworkExtrasProvider mediationNetworkExtrasProvider;
+
+  @SuppressWarnings("deprecation")
+  @Nullable
+  private MediationNetworkExtrasProvider mediationNetworkExtrasProvider;
+
   @NonNull private final FlutterRequestAgentProvider requestAgentProvider;
 
   AdMessageCodec(
@@ -95,6 +98,7 @@ class AdMessageCodec extends StandardMessageCodec {
     this.context = context;
   }
 
+  @SuppressWarnings("deprecation")
   void setMediationNetworkExtrasProvider(
       @Nullable MediationNetworkExtrasProvider mediationNetworkExtrasProvider) {
     this.mediationNetworkExtrasProvider = mediationNetworkExtrasProvider;