+++ packages/google_mobile_ads/lib/src/ump/consent_form_impl.dart
@@ -37,4 +37,7 @@ class ConsentFormImpl extends ConsentForm {
   bool operator ==(Object other) {
     return other is ConsentFormImpl && platformHash == other.platformHash;
   }
+
+  @override
+  int get hashCode => platformHash.hashCode;
 }