+++ packages/mediation/gma_mediation_ironsource/android/src/test/kotlin/io/flutter/plugins/googlemobileads/mediation/gma_mediation_ironsource/GmaMediationIronsourcePluginTest.kt
@@ -1,27 +1,80 @@
+// Copyright 2024 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+// https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 package io.flutter.plugins.googlemobileads.mediation.gma_mediation_ironsource
 
-import io.flutter.plugin.common.MethodCall
-import io.flutter.plugin.common.MethodChannel
-import kotlin.test.Test
-import org.mockito.Mockito
+import androidx.test.ext.junit.runners.AndroidJUnit4
+import com.ironsource.mediationsdk.IronSource
+import io.flutter.plugins.googlemobileads.mediation.gma_mediation_ironsource.GmaMediationIronsourcePlugin.Companion.FALSE
+import io.flutter.plugins.googlemobileads.mediation.gma_mediation_ironsource.GmaMediationIronsourcePlugin.Companion.IRONSOURCE_DONOTSELL_KEY
+import io.flutter.plugins.googlemobileads.mediation.gma_mediation_ironsource.GmaMediationIronsourcePlugin.Companion.TRUE
+import org.junit.Test
+import org.junit.runner.RunWith
+import org.mockito.Mockito.mockStatic
+import org.mockito.kotlin.eq
 
-/*
- * This demonstrates a simple unit test of the Kotlin portion of this plugin's implementation.
- *
- * Once you have built the plugin's example app, you can run these tests from the command
- * line by running `./gradlew testDebugUnitTest` in the `example/android/` directory, or
- * you can run them directly from IDEs that support JUnit such as Android Studio.
- */
-
+@RunWith(AndroidJUnit4::class)
 internal class GmaMediationIronsourcePluginTest {
   @Test
-  fun onMethodCall_getPlatformVersion_returnsExpectedValue() {
+  fun setConsent_withTrueValue_invokesIronSourceSetConsent() {
     val plugin = GmaMediationIronsourcePlugin()
+    mockStatic(IronSource::class.java).use { mockedIronSource ->
 
-    val call = MethodCall("getPlatformVersion", null)
-    val mockResult: MethodChannel.Result = Mockito.mock(MethodChannel.Result::class.java)
-    plugin.onMethodCall(call, mockResult)
+      plugin.setConsent(true)
 
-    Mockito.verify(mockResult).success("Android " + android.os.Build.VERSION.RELEASE)
+      mockedIronSource.verify {
+        IronSource.setConsent(eq(true))
+      }
+    }
+  }
+
+  @Test
+  fun setConsent_withFalseValue_invokesIronSourceSetConsent() {
+    val plugin = GmaMediationIronsourcePlugin()
+    mockStatic(IronSource::class.java).use { mockedIronSource ->
+
+      plugin.setConsent(false)
+
+      mockedIronSource.verify {
+        IronSource.setConsent(eq(false))
+      }
+    }
+  }
+
+  @Test
+  fun setDoNotSell_withTrueValue_invokesIronSourceMetaDataWithCorrectKeyAndValue() {
+    val plugin = GmaMediationIronsourcePlugin()
+    mockStatic(IronSource::class.java).use { mockedIronSource ->
+
+      plugin.setDoNotSell(true)
+
+      mockedIronSource.verify {
+        IronSource.setMetaData(eq(IRONSOURCE_DONOTSELL_KEY), eq(TRUE))
+      }
+    }
+  }
+
+  @Test
+  fun setDoNotSell_withFalseValue_invokesIronSourceMetaDataWithCorrectKeyAndValue() {
+    val plugin = GmaMediationIronsourcePlugin()
+    mockStatic(IronSource::class.java).use { mockedIronSource ->
+
+      plugin.setDoNotSell(false)
+
+      mockedIronSource.verify {
+        IronSource.setMetaData(eq(IRONSOURCE_DONOTSELL_KEY), eq(FALSE))
+      }
+    }
   }
 }