+++ packages/mediation/gma_mediation_applovin/CHANGELOG.md
@@ -1,5 +1,13 @@
-## 1.0.0
+## AppLovin Flutter Mediation Adapter Changelog
 
+#### 1.1.0
+* API for the GmaMediationApplovin is no longer static.
+* Verified compatibility with AppLovin Android adapter version 12.4.3.0.
+* Verified compatibility with AppLovin iOS adapter version 12.4.2.0.
+* Built and tested with the Google Mobile Ads Flutter Plugin version 5.1.0.
+
+#### 1.0.0
 * Initial release.
 * Verified compatibility with AppLovin Android adapter version 12.3.0.0
 * Verified compatibility with AppLovin iOS adapter version 12.2.1.0
+* Built and tested with the Google Mobile Ads Flutter Plugin version 5.0.0.